7727, 7743, 7768 - Adhoc dialog and other dialog-related fixes (#116)
* 7743 Button styles

* 7743 singleInput dialogs

Moved input to its own line, replaced questions with titles, changed button order, changed button labels to both read "Confirm"

* 7743 yesNo Dialog

* 7727 Adhoc Dialog

* 7743 Process Monitor buttons

* 7743 fixed preferences buttons

* 7743 change buttonSize to classsName

This makes the code easier to understand. Buttons are no longer by size.

* 7743 Prune old buttton CSS

All of these properties are now owned by the fsbl-button class. Button size is standardized and no longer used.

* 7768 Authentication Dialog resources

* 7768 Authentication Dialog (read comment)

I applied styles here, and everything looks and feels good. The javascript replacements were the product of a copy-paste, where sales-demo replaced seed.

I tested both sets of javascript, and it behaved well if the sales-demo javascript was in place. I'm assuming this is the javascript we're looking for.

* fixed welcome component image

was the only casualty of an image replacement

* 7768 Restored Seed Auth code

* comment at the top