User now gets more obvious message when renaming a folder to a claimed name
Using this example https://chartiq.kanbanize.com/ctrl_board/18/cards/10467/details we now pop up a confirm box to let the user know we're going to tack a "(1)" onto the name. Unfortunately, using this dialog causes the appLauncher to become unfocused which closes the menu. We'll need to see if we can work around that somehow