hotfix: gulp crashing when e2o is not configured
If e2o is not setup the build process should not break.
This *attempts* to require e2o, but does nothing if not found.
Only if the e2o code is being called do we then throw and error.

This prevents openfin only users from needing to setup e2o.