chore: remove icon support from window tabs
The logo code for tabs has been broken for quite a while.
Upon triaging it to get them back, all kinds of weird state
isolation problems were appearing.

It was decided that rather than address all of this now in 3.x
we will remove what functionality was left of it for this series.
For the ux refresh in 4.x this will be revisited and brought back
in a properly functioning way.

To account for the logo element no longer being in the DOM, which
was giving some space in between the divider and title. We now
have a fixed padding of 8px on the left of the title text.
This prevents the text from directly touching the divider visually.