14689-Dragging-a-tab-can-cause-central-logger-erro
We need to ensure that the key attribute in React
components is unique to that component and will not
change.  This helps React determine whether to
update the DOM or not.

The error that was observed in central logger was a
result of the incorrect tab(s) being updated because
the key was the index in the list of tabs, which
caused React to think that a tab that had been moved
did not need to be re-rendered when in fact it did,
and that a tab that had not been moved needed to be
re-rendered when it fact it did not.

Given the right sequence of actions, this resulted in
componentWillUnmount being called on a component which
had not received a componentWillMount call, hence the
error described in the Kanban ticket.