fix: #id 14512 TilingStop on tabbing cancel (#448)
fix: #id [14512]

Kanban link

Fixed stopTilingOrTabbing
stopTilingOrTabbing had an error in how it was calling window bounds. However, after digging through how this is called and potential ramifications, there was no reason for it to be doing so.

This function has been stripped back. With how this function is called, "allowDropOnSelf" is set to true (via config) from drop events on Finsemble components. With the front end guard being right, we don't need to know if the "pointIsInBox", we just need to know if there's a drop callback to execute.

So, we transmit the router messages, and if the stopTabbingOrTiling event happened on a window that allows it, we run the callback, i.e. the drop event for that component. This should help tighten up the handling of tabs.

Description of testing
Markdown will convert the 1s to the appropriate number.

Open a Welcome Component
Try to drag the tab of a window into it's own body.
Verify that the window can still be moved, tabbed into other windows, etc...
In general, play with tabbing and verify that everything works as designed.