27848 rebranding (#135)
Any instance of the old Finsemble logo has been replaced with the new one.
The primary accent colors were changed to match the new ones.
--primary-accent-2 has been removed. Instances of --primary-accent-2 have been converted to --primary-accent-1. There weren't many of them.
The system tray icon is a modified version of the new logo that has thicker lines. This had to be done otherwise the system tray icon would look dainty and ugly.
Convert splash screen to SVG for high-res screens.  nativeImage.getSize() returns 0,0 for svg, so adjusted to fall back to undefined for defaults